Rename com.paypal.android.ui.vault package to com.paypal.android.ui.vaultcard.
